Move tests from `src/gui/services` to the `tests` directory to centralize all test code and separate it from application logic.
- The test modules within `app_service.rs` and `search_service.rs` have been extracted into their own files within the `tests/gui_services/` directory.
- The `test_helpers.rs` module has been kept in `src` but is now conditionally compiled for tests, making it available to both unit and integration tests in a standard way.
- All relevant import paths have been updated to reflect these changes.
- The code has been formatted with `cargo fmt` and checked with `cargo clippy`.
