+++ tests/gui_services/mod.rs
@@ -1,8 +1,2 @@
-pub mod test_aircraft_service;
-pub mod test_airport_service;
-pub mod test_app_service;
-pub mod test_history_service;
-pub mod test_popup_service;
-pub mod test_route_service;
-pub mod test_search_service;
-pub mod test_validation_service;
+pub mod app_service_tests;
+pub mod search_service_tests;
