+++ tests/lib.rs
@@ -1,5 +1,6 @@
 // Test modules for the flight planner GUI services
 pub mod gui_services;
 
+
 // Test modules for table items
 pub mod table_items_tests;
